refactor: make the Share drawer share-only; 3-dot menu keeps its copy rows
Per review, the shared element is now just the three true Share options
(browser link, image file, image URL) in ShareActionRows, surfaced by the
ShareOptionsBottomSheet drawer.

- The 3-dot menu keeps its four copy-to-clipboard rows inline, exactly as
  before; its share section is now a single "Share" row that opens the drawer
  instead of doing the share directly.
- NoteDropDownMenu owns the drawer state and renders the sheet in place of the
  menu dialog, so the existing direct callers (MultiSetCompose,
  MessageSetCompose) need no changes.
- The reaction-row Share button opens the same drawer.
